+++ feature/home/src/test/kotlin/com/mediahub/feature/home/AuthNavigationPolicyTest.kt
@@ -1,6 +1,5 @@
 package com.mediahub.feature.home
 
-import com.mediahub.model.MediaType
 import com.mediahub.provider.api.AuthSessionErrorKind
 import com.mediahub.provider.api.AuthSessionState
 import org.junit.Assert.assertFalse